Feat: add pagination support (FE-20)
